Move tests in separate files
